拆分GraphView 并将MapSuspense改名为DataSuspense
