+++ src/App.vue
@@ -1,8 +1,9 @@
 <script setup lang="ts">
 import InfoPanel from './components/InfoPanel.vue'
 import VisualPanel from './components/VisualPanel.vue';
+import DataSuspense from './components/DataSuspense.vue';
 import MapView from './components/MapView.vue';
-import MapSuspense from './components/MapSuspense.vue';
+import GraphView from './components/GraphView.vue';
 
 import resInfoIntro from '@/assets/html/intro.html?raw';
 import resInfoFilter from '@/assets/html/filter.html?raw';
@@ -18,9 +19,11 @@ import resIconHeatmap from '@/assets/img/heatmap.png'
     <div name="过滤器"><div v-html="resInfoFilter"/></div>
     <div name="参数"><div v-html="resInfoParam"/></div>
   </InfoPanel>
-  <MapSuspense id="map-view" v-slot="{ graph, map }">
-    <MapView :graph="graph" :map="map"/>
-  </MapSuspense>
+  <DataSuspense id="map-view" v-slot="{ graph, map }">
+    <MapView :map="map" v-slot="{ transform, projection }">
+      <GraphView :graph="graph" :transform="transform" :projection="projection"/>
+    </MapView>
+  </DataSuspense>
   <!-- 右侧栏 -->
   <VisualPanel id="right-panel">
     <div :icon="resIconDensityCurve"></div>
