+++ src/components/DataSuspense.vue
